requirements_app.txt
django~=4.2.1
django-nested-admin~=4.0.2
